refactor: MenuGroup을 Entity로 변경
